Tweak RegisteredMFAMethodListField component
