+++ client/src/components/FormField/RegisteredMFAMethodListField/RegisteredMFAMethodListField.scss
@@ -1,5 +1,5 @@
 .registered-mfa-method-list-field {
-  padding: .5385rem 0;
+  padding: $input-padding-y 0;
 }
 
 .method-list {
